Created Enemy Chase State
We have 3 phases: Chasing, Suspicious and Confirmed.
Always moves to the last known player position.
In the Chasing, check if the player was too close or if the path ended.
In the Suspicious, check if the player was detected or time was ended.
In the Confirmed, check it the times was ended.