Create EnemySearchState
Search State have two phases: Searching and Suspicious.
In the Search phase, moves to the last sound position.
If we reach the sound position, the Suspicious phase starts.
In the Suspicious phase, check for new sounds until the timer ends.
If the timer ends, transition to PatrolState.
If the player is detected (during any phase), transition to ChaseState.