Creation of PlayerArea and PlayerState
A area can be Safe, Normal or Suspicious.
The state can be Normal or Suspicious.
Added methods to set the current Area and State.
Added methods to check if is in a safe area or if is suspicious.
Is considered suspicious if is on a sus Area or have a sus State.
Updated Player prefab.